refactor(api): R2+R3 — split Program.cs into DI modules + partial endpoint files
R2: extract Program.cs (1066->559) service registration into 11
IServiceCollection extension methods (Api/Extensions/ServiceCollectionExtensions.*).
Program reads as a table of contents. Middleware pipeline section byte-identical.

R3: partial-split ReadingTrackingEndpoints (912) and AdminAiQualityEndpoints
(1020) by sub-area, mirroring the VocabularyEndpoints pattern. Entry-point Map
methods + route tables stay in base files; handler bodies moved to partials.

Pure reorganization, zero behavior change. Registration set + route set verified
identical (adversarial QA, all 6 vectors clean). 999 unit tests green.
