+++ backend/src/Api/Endpoints/ReadingTrackingEndpoints.cs
@@ -1,16 +1,23 @@
-using Api.Extensions;
-using Api.Sites;
-using Application.Auth;
 using Application.Common.Interfaces;
-using Application.ReadingTracking;
-using Domain.Entities;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Caching.Memory;
 
 namespace Api.Endpoints;
 
-public static class ReadingTrackingEndpoints
+/// <summary>
+/// Reading-tracking HTTP surface. Routes registered via <see cref="MapReadingTrackingEndpoints"/>;
+/// handlers are split across partial files by sub-domain to keep each file reviewable:
+///
+///   - ReadingTrackingEndpoints.Sessions.cs      SubmitSession, GetSessions
+///   - ReadingTrackingEndpoints.Stats.cs         GetStats, GetDailyStats
+///   - ReadingTrackingEndpoints.Goals.cs         GetGoals, CreateOrUpdateGoal, DeleteGoal
+///   - ReadingTrackingEndpoints.Achievements.cs  GetAchievements
+///   - ReadingTrackingEndpoints.BookStats.cs     GetBookStats
+///   - ReadingTrackingEndpoints.Summary.cs       GetLibrarySummary, GetPace
+///
+/// This file keeps the route table + shared streak/tz helpers + DTOs. Splits use C#
+/// `partial` — compile-identical to the original monolithic file.
+/// </summary>
+public static partial class ReadingTrackingEndpoints
 {
     public static void MapReadingTrackingEndpoints(this WebApplication app)
     {
@@ -27,708 +34,6 @@ public static class ReadingTrackingEndpoints
         group.MapGet("/book-stats", GetBookStats).WithName("GetBookStats");
         group.MapGet("/pace", GetPace).WithName("GetReadingPace");
         group.MapGet("/library-summary", GetLibrarySummary).WithName("GetLibrarySummary");
-    }
-
-    // --- Library Summary (slice 20) ---
-
-    private static async Task<IResult> GetLibrarySummary(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        IMemoryCache cache,
-        [FromQuery] string? tz,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-        var siteId = httpContext.GetSiteId();
-
-        var cacheKey = $"library-summary:{userId.Value}:{siteId}:{tz ?? "0"}";
-        if (cache.TryGetValue<LibrarySummaryDto>(cacheKey, out var cached) && cached != null)
-            return Results.Ok(cached);
-
-        var tzOffset = ParseTzOffset(tz);
-        var now = DateTimeOffset.UtcNow;
-        var todayLocal = GetDayStart(now, tzOffset);
-        var monthStart = new DateTimeOffset(todayLocal.Year, todayLocal.Month, 1, 0, 0, 0, tzOffset).ToUniversalTime();
-        var yearStart = new DateTimeOffset(todayLocal.Year, 1, 1, 0, 0, 0, tzOffset).ToUniversalTime();
-
-        var monthAgg = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.StartedAt >= monthStart)
-            .GroupBy(_ => 1)
-            .Select(g => new { Words = g.Sum(s => (long)s.WordsRead), Seconds = g.Sum(s => (long)s.DurationSeconds) })
-            .FirstOrDefaultAsync(ct);
-
-        var pagesThisMonth = (int)((monthAgg?.Words ?? 0) / 250);
-        var minutesThisMonth = (int)((monthAgg?.Seconds ?? 0) / 60);
-
-        var streakMinMinutes = await GetStreakMinMinutes(db, userId.Value, siteId, ct);
-        var currentStreak = await CalculateStreak(db, userId.Value, siteId, streakMinMinutes, now, ct, tzOffset);
-
-        var booksFinishedYtd = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.EndPercent >= 0.99 && s.EndedAt >= yearStart)
-            .Select(s => s.EditionId ?? s.UserBookId)
-            .Distinct()
-            .CountAsync(ct);
-
-        var dailyGoal = await db.ReadingGoals
-            .Where(g => g.UserId == userId.Value && g.GoalType == "daily_minutes" && g.IsActive)
-            .Select(g => new { g.TargetValue, g.StreakMinMinutes })
-            .FirstOrDefaultAsync(ct);
-        var yearlyGoal = await db.ReadingGoals
-            .Where(g => g.UserId == userId.Value && g.GoalType == "books_per_year" && g.IsActive)
-            .Select(g => new { g.TargetValue, g.Year })
-            .FirstOrDefaultAsync(ct);
-
-        GoalSummaryDto? goal = null;
-        if (dailyGoal != null)
-        {
-            var todayStart = todayLocal.ToUniversalTime();
-            var todaySeconds = await db.ReadingSessions
-                .Where(s => s.UserId == userId.Value && s.StartedAt >= todayStart)
-                .SumAsync(s => (long)s.DurationSeconds, ct);
-            goal = new GoalSummaryDto("daily_minutes", (int)(todaySeconds / 60), dailyGoal.TargetValue);
-        }
-        else if (yearlyGoal != null)
-        {
-            goal = new GoalSummaryDto("books_per_year", booksFinishedYtd, yearlyGoal.TargetValue);
-        }
-
-        var dto = new LibrarySummaryDto(
-            PagesThisMonth: pagesThisMonth,
-            MinutesThisMonth: minutesThisMonth,
-            CurrentStreak: currentStreak,
-            StreakMinMinutes: streakMinMinutes,
-            BooksFinishedYtd: booksFinishedYtd,
-            Goal: goal
-        );
-
-        cache.Set(cacheKey, dto, TimeSpan.FromMinutes(5));
-        return Results.Ok(dto);
-    }
-
-    // --- Reading Pace (slice 19) ---
-
-    private const int PaceMinSessions = 3;
-    private const int FallbackPaceWpm = 200;
-
-    private static async Task<IResult> GetPace(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        IMemoryCache cache,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-        var siteId = httpContext.GetSiteId();
-
-        var cacheKey = $"reading-pace:{userId.Value}:{siteId}";
-        if (cache.TryGetValue<ReadingPaceDto>(cacheKey, out var cached) && cached != null)
-            return Results.Ok(cached);
-
-        var agg = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.WordsRead > 0 && s.DurationSeconds > 0)
-            .GroupBy(_ => 1)
-            .Select(g => new { Sessions = g.Count(), Words = g.Sum(s => (long)s.WordsRead), Seconds = g.Sum(s => (long)s.DurationSeconds) })
-            .FirstOrDefaultAsync(ct);
-
-        ReadingPaceDto dto;
-        if (agg == null || agg.Sessions < PaceMinSessions || agg.Seconds <= 0)
-        {
-            dto = new ReadingPaceDto(FallbackPaceWpm, agg?.Sessions ?? 0, false);
-        }
-        else
-        {
-            var wpm = (int)Math.Round(agg.Words / (agg.Seconds / 60.0));
-            // Clamp to sane range — guards against ultra-short sessions skewing avg
-            wpm = Math.Clamp(wpm, 50, 800);
-            dto = new ReadingPaceDto(wpm, agg.Sessions, true);
-        }
-
-        cache.Set(cacheKey, dto, TimeSpan.FromHours(1));
-        return Results.Ok(dto);
-    }
-
-    // --- Sessions ---
-
-    private static async Task<IResult> SubmitSession(
-        [FromBody] SubmitSessionRequest request,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        ILogger<Program> logger,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-        var siteId = httpContext.GetSiteId();
-
-        // Validation
-        if (request.EditionId == null && request.UserBookId == null)
-            return Results.BadRequest("EditionId or UserBookId required");
-        if (request.DurationSeconds <= 0 || request.DurationSeconds > 14400)
-            return Results.BadRequest("DurationSeconds must be 1-14400");
-        if (request.StartedAt > DateTimeOffset.UtcNow.AddMinutes(5))
-            return Results.BadRequest("StartedAt cannot be in the future");
-        if (request.StartedAt < DateTimeOffset.UtcNow.AddDays(-7))
-            return Results.BadRequest("StartedAt cannot be older than 7 days");
-        var elapsed = (request.EndedAt - request.StartedAt).TotalSeconds;
-        if (elapsed < request.DurationSeconds)
-            return Results.BadRequest("EndedAt - StartedAt must be >= DurationSeconds");
-
-        // Pre-check the unique-by-(user, user_book, started_at) constraint
-        // for user-book sessions. The DB constraint is the source of truth
-        // and the catch below still covers a race; this just keeps the
-        // happy path off the "SQL error level: ERROR" log line that
-        // Npgsql emits before the catch can swallow it. The constraint is
-        // partial (WHERE user_book_id IS NOT NULL), so edition-only
-        // sessions skip the check entirely.
-        if (request.UserBookId != null)
-        {
-            var dup = await db.ReadingSessions.AnyAsync(
-                s => s.UserId == userId.Value
-                  && s.UserBookId == request.UserBookId
-                  && s.StartedAt == request.StartedAt,
-                ct);
-            if (dup) return Results.Ok(new SubmitSessionResponse(Guid.Empty, []));
-        }
-
-        var session = new ReadingSession
-        {
-            Id = Guid.NewGuid(),
-            UserId = userId.Value,
-            SiteId = siteId,
-            EditionId = request.EditionId,
-            UserBookId = request.UserBookId,
-            StartedAt = request.StartedAt,
-            EndedAt = request.EndedAt,
-            DurationSeconds = request.DurationSeconds,
-            WordsRead = request.WordsRead,
-            StartPercent = request.StartPercent,
-            EndPercent = request.EndPercent,
-            CreatedAt = DateTimeOffset.UtcNow,
-        };
-
-        db.ReadingSessions.Add(session);
-
-        try
-        {
-            await db.SaveChangesAsync(ct);
-        }
-        catch (DbUpdateException ex) when (ex.InnerException is Npgsql.PostgresException { SqlState: "23505" })
-        {
-            // Race-window fallback: another concurrent request slipped in
-            // between our pre-check and SaveChanges. The first one won; we
-            // ack idempotently.
-            return Results.Ok(new SubmitSessionResponse(session.Id, []));
-        }
-
-        // Achievement check is best-effort — never fail the session submit because of it.
-        List<string> newAchievements = [];
-        try
-        {
-            var streakMinMinutes = await GetStreakMinMinutes(db, userId.Value, siteId, ct);
-            var currentStreak = await CalculateStreak(db, userId.Value, siteId, streakMinMinutes, request.EndedAt, ct);
-
-            var checker = new AchievementChecker(db);
-            newAchievements = await checker.CheckAfterSession(userId.Value, siteId, session, currentStreak, ct);
-        }
-        catch (Exception ex)
-        {
-            logger.LogWarning(ex, "AchievementChecker failed for session {SessionId}, session still saved", session.Id);
-        }
-
-        return Results.Ok(new SubmitSessionResponse(session.Id, newAchievements));
-    }
-
-    private static async Task<IResult> GetSessions(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] DateTimeOffset? from,
-        [FromQuery] DateTimeOffset? to,
-        [FromQuery] int? limit,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        var query = db.ReadingSessions
-            .Where(s => s.UserId == userId.Value);
-
-        if (from.HasValue) query = query.Where(s => s.StartedAt >= from.Value);
-        if (to.HasValue) query = query.Where(s => s.StartedAt <= to.Value);
-
-        var sessions = await query
-            .OrderByDescending(s => s.StartedAt)
-            .Take(limit ?? 100)
-            .Select(s => new SessionDto(
-                s.Id, s.EditionId, s.UserBookId,
-                s.StartedAt, s.EndedAt, s.DurationSeconds,
-                s.WordsRead, s.StartPercent, s.EndPercent))
-            .ToListAsync(ct);
-
-        return Results.Ok(sessions);
-    }
-
-    // --- Stats ---
-
-    private static async Task<IResult> GetStats(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] string? tz,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-        var siteId = httpContext.GetSiteId();
-
-        var tzOffset = ParseTzOffset(tz);
-        var now = DateTimeOffset.UtcNow;
-
-        var allSessions = db.ReadingSessions
-            .Where(s => s.UserId == userId.Value);
-
-        var totalSeconds = await allSessions.SumAsync(s => (long)s.DurationSeconds, ct);
-        var totalWords = await allSessions.SumAsync(s => (long)s.WordsRead, ct);
-        var sessionCount = await allSessions.CountAsync(ct);
-
-        var booksFinished = await allSessions
-            .Where(s => s.EndPercent >= 0.99)
-            .Select(s => s.EditionId ?? s.UserBookId)
-            .Distinct()
-            .CountAsync(ct);
-
-        // Time-period sums (calculate in user tz, convert to UTC for PostgreSQL)
-        var todayLocal = GetDayStart(now, tzOffset);
-        var todayStart = todayLocal.ToUniversalTime();
-        var weekStart = todayLocal.AddDays(-(int)todayLocal.DayOfWeek).ToUniversalTime();
-        var monthStart = new DateTimeOffset(todayLocal.Year, todayLocal.Month, 1, 0, 0, 0, tzOffset).ToUniversalTime();
-
-        var todaySeconds = await allSessions
-            .Where(s => s.StartedAt >= todayStart)
-            .SumAsync(s => (long)s.DurationSeconds, ct);
-        var weekSeconds = await allSessions
-            .Where(s => s.StartedAt >= weekStart)
-            .SumAsync(s => (long)s.DurationSeconds, ct);
-        var monthSeconds = await allSessions
-            .Where(s => s.StartedAt >= monthStart)
-            .SumAsync(s => (long)s.DurationSeconds, ct);
-
-        // Streak
-        var streakMinMinutes = await GetStreakMinMinutes(db, userId.Value, siteId, ct);
-        var currentStreak = await CalculateStreak(db, userId.Value, siteId, streakMinMinutes, now, ct, tzOffset);
-        var longestStreak = await CalculateLongestStreak(db, userId.Value, siteId, streakMinMinutes, ct, tzOffset);
-
-        // Averages
-        double avgDailyMinutes = 0;
-        double avgWordsPerMinute = 0;
-        if (sessionCount > 0)
-        {
-            var firstSession = await allSessions
-                .OrderBy(s => s.StartedAt)
-                .Select(s => s.StartedAt)
-                .FirstOrDefaultAsync(ct);
-            var daysSinceFirst = Math.Max(1, (now - firstSession).TotalDays);
-            avgDailyMinutes = totalSeconds / 60.0 / daysSinceFirst;
-
-            if (totalSeconds > 0)
-                avgWordsPerMinute = totalWords / (totalSeconds / 60.0);
-        }
-
-        // Vocab reviews today
-        var todayVocabReviews = await db.VocabularyReviews
-            .Where(r => r.UserId == userId.Value && r.CreatedAt >= todayStart)
-            .CountAsync(ct);
-
-        // Daily goal (reading + vocab reviews as effective minutes)
-        var dailyGoal = await db.ReadingGoals
-            .Where(g => g.UserId == userId.Value && g.GoalType == "daily_minutes" && g.IsActive)
-            .FirstOrDefaultAsync(ct);
-
-        object? dailyGoalObj = null;
-        if (dailyGoal != null)
-        {
-            var effectiveTodayMinutes = todaySeconds / 60.0 + todayVocabReviews * 0.5;
-            dailyGoalObj = new
-            {
-                target = dailyGoal.TargetValue,
-                today = Math.Round(effectiveTodayMinutes, 1),
-                met = effectiveTodayMinutes >= dailyGoal.TargetValue,
-            };
-        }
-
-        return Results.Ok(new
-        {
-            totalSeconds,
-            totalWords,
-            booksFinished,
-            currentStreak,
-            longestStreak,
-            streakMinMinutes,
-            avgDailyMinutes = Math.Round(avgDailyMinutes, 1),
-            avgWordsPerMinute = Math.Round(avgWordsPerMinute, 1),
-            todaySeconds,
-            todayVocabReviews,
-            weekSeconds,
-            monthSeconds,
-            dailyGoal = dailyGoalObj,
-        });
-    }
-
-    private static async Task<IResult> GetDailyStats(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] DateTimeOffset? from,
-        [FromQuery] DateTimeOffset? to,
-        [FromQuery] string? tz,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        var tzOffset = ParseTzOffset(tz);
-        var now = DateTimeOffset.UtcNow;
-        var start = from ?? now.AddDays(-90);
-        var end = to ?? now;
-
-        // Get raw sessions in range
-        var sessions = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.StartedAt >= start && s.StartedAt <= end)
-            .Select(s => new { s.StartedAt, s.DurationSeconds, s.WordsRead })
-            .ToListAsync(ct);
-
-        // Group by local date
-        var daily = sessions
-            .GroupBy(s => s.StartedAt.ToOffset(tzOffset).Date)
-            .Select(g => new DailyStatDto(
-                g.Key,
-                g.Sum(s => s.DurationSeconds),
-                g.Sum(s => s.WordsRead),
-                g.Count()))
-            .OrderBy(d => d.Date)
-            .ToList();
-
-        return Results.Ok(daily);
-    }
-
-    // --- Goals ---
-
-    private static async Task<IResult> GetGoals(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        var goals = await db.ReadingGoals
-            .Where(g => g.UserId == userId.Value && g.IsActive)
-            .Select(g => new GoalDto(g.Id, g.GoalType, g.TargetValue, g.Year, g.StreakMinMinutes, g.UpdatedAt))
-            .ToListAsync(ct);
-
-        return Results.Ok(goals);
-    }
-
-    private static async Task<IResult> CreateOrUpdateGoal(
-        [FromBody] CreateGoalRequest request,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-        var siteId = httpContext.GetSiteId();
-
-        if (request.GoalType != "daily_minutes" && request.GoalType != "books_per_year")
-            return Results.BadRequest("GoalType must be daily_minutes or books_per_year");
-        if (request.TargetValue <= 0)
-            return Results.BadRequest("TargetValue must be positive");
-
-        var existing = await db.ReadingGoals
-            .FirstOrDefaultAsync(g => g.UserId == userId.Value && g.GoalType == request.GoalType, ct);
-
-        if (existing != null)
-        {
-            existing.TargetValue = request.TargetValue;
-            existing.Year = request.Year;
-            existing.IsActive = true;
-            if (request.StreakMinMinutes.HasValue)
-                existing.StreakMinMinutes = request.StreakMinMinutes.Value;
-            existing.UpdatedAt = DateTimeOffset.UtcNow;
-        }
-        else
-        {
-            existing = new ReadingGoal
-            {
-                Id = Guid.NewGuid(),
-                UserId = userId.Value,
-                SiteId = siteId,
-                GoalType = request.GoalType,
-                TargetValue = request.TargetValue,
-                Year = request.Year,
-                IsActive = true,
-                StreakMinMinutes = request.StreakMinMinutes ?? 5,
-                CreatedAt = DateTimeOffset.UtcNow,
-                UpdatedAt = DateTimeOffset.UtcNow,
-            };
-            db.ReadingGoals.Add(existing);
-        }
-
-        await db.SaveChangesAsync(ct);
-
-        return Results.Ok(new GoalDto(existing.Id, existing.GoalType, existing.TargetValue,
-            existing.Year, existing.StreakMinMinutes, existing.UpdatedAt));
-    }
-
-    private static async Task<IResult> DeleteGoal(
-        Guid id,
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        var goal = await db.ReadingGoals
-            .FirstOrDefaultAsync(g => g.Id == id && g.UserId == userId.Value, ct);
-        if (goal == null) return Results.NotFound();
-
-        db.ReadingGoals.Remove(goal);
-        await db.SaveChangesAsync(ct);
-
-        return Results.NoContent();
-    }
-
-    // --- Achievements ---
-
-    private static async Task<IResult> GetAchievements(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        var unlocked = await db.UserAchievements
-            .Where(a => a.UserId == userId.Value)
-            .Select(a => new AchievementDto(a.AchievementCode, a.UnlockedAt))
-            .ToListAsync(ct);
-
-        return Results.Ok(unlocked);
-    }
-
-    // --- Book Stats ---
-
-    private static async Task<IResult> GetBookStats(
-        HttpContext httpContext,
-        AuthService authService,
-        IAppDbContext db,
-        [FromQuery] int? year,
-        CancellationToken ct)
-    {
-        var userId = httpContext.GetUserId(authService);
-        if (userId == null) return Results.Unauthorized();
-
-        // 1. Find finished editions: first session where EndPercent >= 0.99 per edition
-        var finishEvents = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.EditionId != null && s.EndPercent >= 0.99)
-            .GroupBy(s => s.EditionId!.Value)
-            .Select(g => new { EditionId = g.Key, FinishedAt = g.Min(s => s.EndedAt) })
-            .ToListAsync(ct);
-
-        // Also find finished user books (with metadata for breakdowns)
-        var userBookFinishEvents = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.UserBookId != null && s.EndPercent >= 0.99)
-            .GroupBy(s => s.UserBookId!.Value)
-            .Select(g => new { UserBookId = g.Key, FinishedAt = g.Min(s => s.EndedAt) })
-            .ToListAsync(ct);
-
-        // Available years (both editions and user books)
-        var availableYears = finishEvents.Select(f => f.FinishedAt.Year)
-            .Concat(userBookFinishEvents.Select(f => f.FinishedAt.Year))
-            .Distinct()
-            .OrderDescending()
-            .ToList();
-
-        // 2. Year filter
-        if (year.HasValue && year.Value > 0)
-        {
-            finishEvents = finishEvents.Where(f => f.FinishedAt.Year == year.Value).ToList();
-            userBookFinishEvents = userBookFinishEvents.Where(f => f.FinishedAt.Year == year.Value).ToList();
-        }
-
-        var editionIds = finishEvents.Select(f => f.EditionId).ToList();
-        var finishMap = finishEvents.ToDictionary(f => f.EditionId, f => f.FinishedAt);
-        var userBookIds = userBookFinishEvents.Select(f => f.UserBookId).ToList();
-        var userBookFinishMap = userBookFinishEvents.ToDictionary(f => f.UserBookId, f => f.FinishedAt);
-
-        // 3. Load edition metadata
-        var editions = await db.Editions
-            .Where(e => editionIds.Contains(e.Id))
-            .Select(e => new
-            {
-                e.Id,
-                e.Language,
-                WordCount = e.Chapters.Sum(c => (int?)c.WordCount) ?? 0,
-                Genres = e.Genres.Select(g => new { g.Name, g.Slug }).ToList(),
-                Authors = e.EditionAuthors
-                    .Where(ea => ea.Role == Domain.Enums.AuthorRole.Author)
-                    .OrderBy(ea => ea.Order)
-                    .Select(ea => new { ea.Author.Name, ea.Author.Slug })
-                    .ToList(),
-            })
-            .ToListAsync(ct);
-
-        // 3b. Load finished user book metadata
-        var userBooks = userBookIds.Count > 0
-            ? await db.UserBooks
-                .Where(b => userBookIds.Contains(b.Id))
-                .Select(b => new { b.Id, b.Language, b.Genre, b.Author, b.TotalWordCount })
-                .ToListAsync(ct)
-            : [];
-
-        // 4. First session per edition (for avg days to finish)
-        var firstSessions = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.EditionId != null && editionIds.Contains(s.EditionId!.Value))
-            .GroupBy(s => s.EditionId!.Value)
-            .Select(g => new { EditionId = g.Key, FirstStarted = g.Min(s => s.StartedAt) })
-            .ToListAsync(ct);
-        var firstSessionMap = firstSessions.ToDictionary(f => f.EditionId, f => f.FirstStarted);
-
-        // 5. Reading time per edition (for pace + time-by-genre/author)
-        var sessionsByEdition = await db.ReadingSessions
-            .Where(s => s.UserId == userId.Value && s.EditionId != null && editionIds.Contains(s.EditionId!.Value))
-            .GroupBy(s => s.EditionId!.Value)
-            .Select(g => new { EditionId = g.Key, TotalSeconds = g.Sum(s => s.DurationSeconds), TotalWords = g.Sum(s => s.WordsRead) })
-            .ToListAsync(ct);
-        var sessionMap = sessionsByEdition.ToDictionary(s => s.EditionId);
-
-        // --- Aggregations ---
-        var userBookPages = userBooks.Sum(b => (b.TotalWordCount ?? 0)) / 250;
-        var totalPages = editions.Sum(e => e.WordCount) / 250 + userBookPages;
-
-        // Avg days to finish
-        var daysToFinish = editions
-            .Where(e => finishMap.ContainsKey(e.Id) && firstSessionMap.ContainsKey(e.Id))
-            .Select(e => (finishMap[e.Id] - firstSessionMap[e.Id]).TotalDays)
-            .ToList();
-        var avgDaysToFinish = daysToFinish.Count > 0 ? Math.Round(daysToFinish.Average(), 1) : 0;
-
-        // Genre stats (editions + user books with genre)
-        var genreItems = editions
-            .SelectMany(e => e.Genres.Select(g => new { g.Name, g.Slug }))
-            .Concat(userBooks
-                .Where(b => !string.IsNullOrEmpty(b.Genre))
-                .Select(b => new { Name = b.Genre!, Slug = b.Genre!.ToLowerInvariant().Replace(' ', '-') }));
-        var genreStats = genreItems
-            .GroupBy(g => g.Slug)
-            .Select(g => new GenreStatDto(g.First().Name, g.Key, g.Count()))
-            .OrderByDescending(g => g.Count)
-            .ToList();
-
-        // Author stats (editions + user books with author)
-        var authorItems = editions
-            .SelectMany(e => e.Authors.Select(a => new { a.Name, a.Slug }))
-            .Concat(userBooks
-                .Where(b => !string.IsNullOrEmpty(b.Author))
-                .Select(b => new { Name = b.Author!, Slug = b.Author!.ToLowerInvariant().Replace(' ', '-') }));
-        var authorStats = authorItems
-            .GroupBy(a => a.Slug)
-            .Select(g => new AuthorStatDto(g.First().Name, g.Key, g.Count()))
-            .OrderByDescending(a => a.Count)
-            .Take(10)
-            .ToList();
-
-        // Language stats (editions + user books)
-        var languageItems = editions.Select(e => e.Language)
-            .Concat(userBooks.Select(b => b.Language));
-        var languageStats = languageItems
-            .GroupBy(l => l)
-            .Select(g => new LanguageStatDto(g.Key, g.Count()))
-            .OrderByDescending(l => l.Count)
-            .ToList();
-
-        // Books over time (editions + user books)
-        var allFinishEvents = finishEvents
-            .Select(f => new { f.FinishedAt, EditionId = (Guid?)f.EditionId, UserBookId = (Guid?)null })
-            .Concat(userBookFinishEvents
-                .Select(f => new { f.FinishedAt, EditionId = (Guid?)null, UserBookId = (Guid?)f.UserBookId }));
-        var booksOverTime = allFinishEvents
-            .GroupBy(f => year.HasValue && year.Value > 0
-                ? f.FinishedAt.ToString("yyyy-MM")
-                : f.FinishedAt.Year.ToString())
-            .Select(g =>
-            {
-                var eIds = g.Where(f => f.EditionId != null).Select(f => f.EditionId!.Value).ToHashSet();
-                var ubIds = g.Where(f => f.UserBookId != null).Select(f => f.UserBookId!.Value).ToHashSet();
-                var pages = editions.Where(e => eIds.Contains(e.Id)).Sum(e => e.WordCount) / 250
-                          + userBooks.Where(b => ubIds.Contains(b.Id)).Sum(b => (b.TotalWordCount ?? 0)) / 250;
-                return new BooksOverTimeDto(g.Key, g.Count(), pages);
-            })
-            .OrderBy(b => b.Period)
-            .ToList();
-
-        // Book length distribution (editions + user books)
-        var lengthPages = editions.Select(e => e.WordCount / 250)
-            .Concat(userBooks.Where(b => b.TotalWordCount > 0).Select(b => b.TotalWordCount!.Value / 250));
-        var lengthBuckets = lengthPages
-            .GroupBy(p => p < 150 ? "short" : p < 400 ? "medium" : "long")
-            .Select(g => new BookLengthBucketDto(g.Key, g.Count()))
-            .ToList();
-
-        // Pace stats (auto-calc: seconds per word → slow/medium/fast)
-        var paceList = editions
-            .Where(e => sessionMap.ContainsKey(e.Id) && e.WordCount > 0)
-            .Select(e =>
-            {
-                var wpm = sessionMap[e.Id].TotalWords / Math.Max(1.0, sessionMap[e.Id].TotalSeconds / 60.0);
-                return wpm < 150 ? "slow" : wpm < 300 ? "medium" : "fast";
-            })
-            .GroupBy(p => p)
-            .Select(g => new PaceStatDto(g.Key, g.Count()))
-            .ToList();
-
-        // Reading time by genre
-        var timeByGenre = editions
-            .Where(e => sessionMap.ContainsKey(e.Id))
-            .SelectMany(e => e.Genres.Select(g => new { g.Name, g.Slug, sessionMap[e.Id].TotalSeconds }))
-            .GroupBy(x => x.Slug)
-            .Select(g => new ReadingTimeStatDto(g.First().Name, g.Key, g.Sum(x => x.TotalSeconds)))
-            .OrderByDescending(x => x.Seconds)
-            .ToList();
-
-        // Reading time by author
-        var timeByAuthor = editions
-            .Where(e => sessionMap.ContainsKey(e.Id))
-            .SelectMany(e => e.Authors.Select(a => new { a.Name, a.Slug, sessionMap[e.Id].TotalSeconds }))
-            .GroupBy(x => x.Slug)
-            .Select(g => new ReadingTimeStatDto(g.First().Name, g.Key, g.Sum(x => x.TotalSeconds)))
-            .OrderByDescending(x => x.Seconds)
-            .Take(10)
-            .ToList();
-
-        return Results.Ok(new BookStatsResponse(
-            BooksFinished: editionIds.Count + userBookIds.Count,
-            TotalPages: totalPages,
-            AvgDaysToFinish: avgDaysToFinish,
-            GenreStats: genreStats,
-            AuthorStats: authorStats,
-            LanguageStats: languageStats,
-            BooksOverTime: booksOverTime,
-            BookLengthDistribution: lengthBuckets,
-            PaceStats: paceList,
-            ReadingTimeByGenre: timeByGenre,
-            ReadingTimeByAuthor: timeByAuthor,
-            AvailableYears: availableYears
-        ));
     }
 
     // --- Helpers ---
