+++ backend/src/Api/Program.cs
@@ -1,557 +1,50 @@
-using System.Threading.RateLimiting;
 using Api.Endpoints;
+using Api.Extensions;
 using Api.Language;
 using Api.Middleware;
 using Api.Sites;
 using Application;
 using Application.AdminAuth;
 using Application.Common.Interfaces;
+using Application.Search;
 using Application.TextStack;
 using Domain.Enums;
 using Infrastructure.Persistence;
-using Infrastructure.Services;
-using Infrastructure.Telemetry;
 using Microsoft.AspNetCore.HttpOverrides;
-using Microsoft.AspNetCore.RateLimiting;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Diagnostics;
 using Microsoft.Extensions.FileProviders;
-using Npgsql;
-using TextStack.Ai.Rag;
-using TextStack.Search;
-using TextStack.Search.Abstractions;
-using TextStack.Search.Meilisearch;
-using TextStack.Tts;
-using TextStack.Vocabulary;
-using Api.Services;
-using Application.Auth;
-using Application.Search;
-using OpenTelemetry.Trace;
 using Scalar.AspNetCore;
-using TextStack.Ai.Tools;
 
 var builder = WebApplication.CreateBuilder(args);
 
-// Match nginx client_max_body_size (500MB) + per-user storage quota.
-// Default Kestrel cap is 30MB → users hit 413 on real-world PDFs.
-const long MaxUploadBytes = 500L * 1024 * 1024;
-builder.WebHost.ConfigureKestrel(o => o.Limits.MaxRequestBodySize = MaxUploadBytes);
-builder.Services.Configure<Microsoft.AspNetCore.Http.Features.FormOptions>(o =>
-{
-    o.MultipartBodyLengthLimit = MaxUploadBytes;
-    o.ValueLengthLimit = int.MaxValue;
-    o.MultipartHeadersLengthLimit = int.MaxValue;
-});
+// Host-level configuration (Kestrel/multipart limits + OpenTelemetry).
+builder.AddTextStackUploadLimits();
+builder.AddTextStackObservability();
 
-// OpenTelemetry
-builder.Services.AddTextStackTelemetry(
-    builder.Configuration,
-    "textstack-api",
-    tracing => tracing
-        .AddAspNetCoreInstrumentation(options =>
-        {
-            options.RecordException = true;
-            options.EnrichWithHttpRequest = (activity, request) =>
-            {
-                activity.SetTag("http.client_ip", request.HttpContext.Connection.RemoteIpAddress?.ToString());
-            };
-        })
-        .AddHttpClientInstrumentation());
-builder.Logging.AddTelemetryLogging(builder.Configuration, "textstack-api");
-
-var corsOrigins = builder.Configuration.GetSection("Cors:AllowedOrigins").Get<string[]>()
-    ?? [
-        "http://localhost:5173", "http://general.localhost", "http://general.localhost:5173",
-        "http://localhost:81", "http://admin.localhost", "http://admin.localhost:81",
-        "https://textstack.app", "https://textstack.dev"
-    ];
-builder.Services.AddCors(options =>
-{
-    options.AddDefaultPolicy(policy =>
-    {
-        policy.WithOrigins(corsOrigins)
-            .AllowAnyHeader()
-            .AllowAnyMethod()
-            .AllowCredentials();
-    });
-});
-
+builder.Services.AddTextStackCors(builder.Configuration);
 builder.Services.AddOpenApi();
 
-// Application layer
-builder.Services.AddApplication();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.EvalSuiteRunner>();
-// Scheduled-eval support (Phase 12 RLOps slice 5a): shared single-slot overlap gate +
-// pure regression detector, consumed by both the admin trigger and ContinuousEvalWorker.
-builder.Services.AddSingleton<Application.Ai.IEvalRunGate, Application.Ai.EvalRunGate>();
-builder.Services.AddSingleton<Application.Ai.EvalRegressionDetector>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.RagEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.UserBookRagEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.ToolCallEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.StudyBuddyEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.EnrichmentEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.LibrarianEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.TutorEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.CriticDefectEvalRunner>();
-builder.Services.AddSingleton<TextStack.Ai.EvalSuite.CrewAbEvalRunner>();
-// Tool catalogue (AI-029/030): scans Application for ITool impls; dispatch is schema-validated.
-builder.Services.AddAiTools(typeof(Application.Tools.GetChapterTool).Assembly);
-// Agent loop engine (Phase 6, AI-034). Concrete agents (StudyBuddy, AI-035) build on it.
-TextStack.Ai.Agents.ServiceCollectionExtensions.AddAiAgents(builder.Services);
-builder.Services.AddScoped<Application.Agents.StudyBuddyAgent>();
-// Enrichment agent (AI-Agent-1): registered in the API too so the admin eval path can run it.
-builder.Services.AddScoped<Application.Agents.EnrichmentAgent>();
-// Librarian agent (AI-Agent-3): NL catalog request → ranked, reasoned recommendations. Scoped (its
-// search tools resolve the scoped IAppDbContext + LibrarySearchService per request).
-builder.Services.AddScoped<Application.Search.LibrarySearchService>();
-builder.Services.AddScoped<Application.Agents.LibrarianAgent>();
-// Learning Tutor agent (AI-Agent-2): plans an ordered study set over the learner's SRS + reading state and
-// hands off to the existing vocabulary-review flow. Scoped (its tools resolve the scoped IAppDbContext +
-// IRagService per request).
-builder.Services.AddScoped<Application.Agents.TutorAgent>();
-// Crew specialists (Phase 7, AI-041): single-call IAgent<TIn,TOut> sub-agents the content crews
-// (AI-042/043) compose via CrewTasks.Of. Stateless + ILlmService is a singleton, so singleton is fine.
-builder.Services.AddSingleton<Application.Agents.ResearcherAgent>();
-builder.Services.AddSingleton<Application.Agents.DrafterAgent>();
-builder.Services.AddSingleton<Application.Agents.CriticAgent>();
-builder.Services.AddSingleton<Application.Agents.EditorAgent>();
-// Single-call A/B baseline (Phase 7, AI-046): the "A" arm of the crew-vs-single-call eval — one ILlmService
-// call that writes a field directly under the crew's full brief contract. Stateless singleton like the specialists.
-builder.Services.AddSingleton<Application.Agents.BaselineFieldAgent>();
-// AutoPublish crew (Phase 7, AI-042): in-process admin path that runs the specialists over ILlmService to
-// generate SEO prose for an Edition. Scoped because it persists via the scoped IAgentRunWriter (per-request
-// DbContext). The legacy bash + Claude-CLI poller stays the default; this is the observable, traced alternative.
-// Shared per-field crew runner (Phase 7, AI-043): the 4-stage plan + fail-closed gate + agent_run persistence,
-// extracted from AutoPublishCrew so both AutoPublish and SEO crews delegate to it. Scoped (scoped IAgentRunWriter).
-builder.Services.AddScoped<Application.Agents.FieldCrew>();
-builder.Services.AddScoped<Application.Agents.AutoPublishCrew>();
-// SEO crew (Phase 7, AI-043): in-process admin path that generates ONE SEO prose field over ILlmService and
-// routes it through the SeoBackfillJob apply path. The legacy seo-backfill poller stays the default + untouched.
-builder.Services.AddScoped<Application.Agents.SeoCrew>();
-builder.Services.AddAuthSettings(builder.Configuration);
+// Application layer + AI platform (eval runners, agents, tools, crews).
+builder.Services
+    .AddApplication()
+    .AddTextStackAiEvals()
+    .AddTextStackAgents()
+    .AddAuthSettings(builder.Configuration);
 
 var connectionString = builder.Configuration.GetConnectionString("Default")
     ?? throw new InvalidOperationException("ConnectionStrings:Default is required");
 
-builder.Services.AddSingleton<Application.Common.Interfaces.ICurrentSite>(
-    sp => new Infrastructure.Persistence.CurrentSite(sp.GetRequiredService<IConfiguration>()));
-
-builder.Services.AddDbContext<AppDbContext>(options =>
-    options.UseNpgsql(connectionString, o => o.UseVector())
-        .UseSnakeCaseNamingConvention()
-        .ConfigureWarnings(w => w.Ignore(RelationalEventId.PendingModelChangesWarning)));
-
-builder.Services.AddScoped<IAppDbContext>(sp => sp.GetRequiredService<AppDbContext>());
-
-// File storage
+// storagePath is consumed again by the static-file middleware below.
 var storagePath = builder.Configuration["Storage:RootPath"] ?? "/storage";
-builder.Services.AddSingleton<IFileStorageService>(new LocalFileStorageService(storagePath));
 
-// Search library
-builder.Services.AddTextStackSearch();
-var searchProvider = builder.Configuration["Search:Provider"] ?? "postgres";
-if (searchProvider == "meilisearch")
-    builder.Services.AddMeilisearchProvider(options =>
-        builder.Configuration.GetSection("Search:Meilisearch").Bind(options));
-else
-    builder.Services.AddPostgresFtsProvider(
-        _ => () => new NpgsqlConnection(connectionString),
-        options => options.ConnectionString = connectionString);
-
-// RAG vector retrieval (Phase 4). Raw Npgsql like the search provider; query embedded via
-// IEmbeddingService (registered in AddApplication). Used by the admin debug endpoint (AI-022).
-builder.Services.AddRagRetrieval(_ => () => new NpgsqlConnection(connectionString));
-// On-demand "Ask this book" index trigger (Phase 1): the chunker + shared chunking service.
-builder.Services.AddAiRag();
-builder.Services.AddScoped<Infrastructure.Rag.BookChunkingService>();
-// Spoiler-safe context builder (AI-024): resolves lastRead + gates chunks + private corpus.
-builder.Services.AddScoped<Application.Rag.RagContextService>();
-// "Ask this book" orchestration (AI-025): context + LLM gateway → grounded answer with citations.
-builder.Services.AddScoped<Application.Rag.RagAskService>();
-builder.Services.AddScoped<Application.Rag.IRagAskService>(sp => sp.GetRequiredService<Application.Rag.RagAskService>());
-// Phase 2 "Ask this book" for user uploads: per-user isolated retrieval context (no spoiler gate).
-builder.Services.AddScoped<Application.Rag.UserBookRagContextService>();
-
-// Similar books (AI-055): cosine NN over editions.embedding. Same raw Func<IDbConnection>
-// factory as RAG — a raw connection casts the vector server-side (no pgvector type registration).
-builder.Services.AddScoped(_ =>
-    new Application.Recommendations.SimilarBooksService(() => new NpgsqlConnection(connectionString)));
-
-// Hybrid catalog search (AI-057): blends the FTS edition ranking with cosine NN over
-// editions.embedding via RRF. Only invoked on `semantic=true`; the pure-FTS path never touches it.
-builder.Services.AddScoped(sp =>
-    new Application.Search.HybridCatalogSearch(
-        sp.GetRequiredService<TextStack.Search.Abstractions.ISearchProvider>(),
-        // Lazy: OpenAiEmbeddingClient throws in its ctor on a keyless host. Resolved only when
-        // semantic search actually runs (inside HybridCatalogSearch's try/catch), so non-semantic
-        // /search never constructs it and a keyless stack degrades to FTS instead of 500ing.
-        () => sp.GetRequiredService<global::TextStack.Ai.Core.IEmbeddingService>(),
-        () => new NpgsqlConnection(connectionString),
-        sp.GetRequiredService<ILogger<Application.Search.HybridCatalogSearch>>()));
-
-// Reindex service (used by CLI)
-builder.Services.AddScoped<SearchReindexService>();
-
-// Vocabulary SRS library
-builder.Services.AddTextStackVocabulary(options =>
-{
-    options.OllamaBaseUrl = builder.Configuration["Ollama:BaseUrl"] ?? "http://localhost:11434";
-    options.OllamaModel = builder.Configuration["Ollama:Model"] ?? "gemma4:e2b";
-    options.OllamaTimeoutSeconds = builder.Configuration.GetValue("Ollama:TimeoutSeconds", 30);
-});
-
-// Image optimization
-builder.Services.AddSingleton<IImageOptimizer, ImageOptimizer>();
-
-// Site resolution
-builder.Services.AddMemoryCache();
-builder.Services.AddSingleton<ISiteResolver, SiteResolver>();
-
-// TextStack import
-builder.Services.AddScoped<TextStackImportService>();
-
-// User books
-builder.Services.AddScoped<Application.UserBooks.UserBookService>();
-builder.Services.AddScoped<Application.UserBooks.MetadataService>();
-builder.Services.AddScoped<Application.UserBooks.TagService>();
-builder.Services.AddScoped<Application.Collections.CollectionService>();
-builder.Services.AddScoped<Application.UserBooks.BulkActionService>();
-builder.Services.AddScoped<Application.UserBooks.BookStatsService>();
-builder.Services.AddScoped<Application.UserBooks.UserBookSearchService>();
-builder.Services.AddScoped<Application.Library.LibraryShelvesService>();
-
-// Standard Ebooks sync
-builder.Services.AddHttpClient<StandardEbooksSyncService>();
-builder.Services.AddScoped<StandardEbooksSyncService>();
-
-// Email (Resend)
-builder.Services.Configure<EmailSettings>(options =>
-{
-    options.ResendApiKey = builder.Configuration["Resend:ApiKey"] ?? "";
-    options.FromEmail = builder.Configuration["Resend:FromEmail"] ?? "noreply@textstack.app";
-    options.BaseUrl = builder.Configuration["App:BaseUrl"] ?? "https://textstack.app";
-    options.AdminAlertEmail = builder.Configuration["Resend:AdminAlertEmail"] ?? "";
-});
-builder.Services.AddHttpClient<IEmailService, ResendEmailService>();
-
-// HttpClient for translation proxy
-builder.Services.AddHttpClient();
-
-// TTS
-builder.Services.Configure<TtsConfiguration>(builder.Configuration.GetSection("Tts"));
-builder.Services.AddSingleton<ITtsService, EdgeTtsService>();
-builder.Services.AddHostedService(sp => (EdgeTtsService)sp.GetRequiredService<ITtsService>());
-
-
-// SSG periodic rebuild
-builder.Services.AddHostedService<SsgPeriodicRebuildWorker>();
-
-// Vocabulary anti-spiral: periodic auto-retire sweep (F4)
-builder.Services.AddHostedService<AutoRetireSweeperWorker>();
-
-// Vocabulary anti-spiral: daily-cap pending → SRS reconciler (F2)
-builder.Services.AddHostedService<DailyCapReconcilerWorker>();
-
-// Vocabulary anti-spiral: one-shot seed of word_frequencies from embedded gz (F1)
-builder.Services.AddHostedService<WordFrequencyLoaderWorker>();
-
-// Vocabulary anti-spiral: 24h cluster candidate builder (F3)
-builder.Services.AddHostedService<ClusterCandidateBuilderWorker>();
-
-// AI-058: weekly semantic concept clustering (groups vocab by meaning across all books)
-builder.Services.AddHostedService<ConceptClusteringWorker>();
-// Phase 12 RLOps slice 5a: scheduled continuous evals (OFF by default — Eval:Scheduled:Enabled).
-builder.Services.AddHostedService<ContinuousEvalWorker>();
-// Phase 12 RLOps slice 5b: embedding-drift detection (OFF by default — Drift:Enabled).
-builder.Services.AddHostedService<DriftDetectionWorker>();
-
-// Rate limiting
-builder.Services.AddRateLimiter(options =>
-{
-    options.AddFixedWindowLimiter("admin-login", opt =>
-    {
-        opt.Window = TimeSpan.FromMinutes(1);
-        opt.PermitLimit = 5;
-        opt.QueueLimit = 0;
-    });
-    options.AddFixedWindowLimiter("user-login", opt =>
-    {
-        opt.Window = TimeSpan.FromMinutes(1);
-        opt.PermitLimit = 10;
-        opt.QueueLimit = 0;
-    });
-    // Device Authorization Grant (RFC 8628, AI-050a) — all per-IP.
-    // device-code: CLI requests a device_code; one per CLI session, 5/min is ample.
-    options.AddPolicy("device-code", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 5,
-            QueueLimit = 0,
-        });
-    });
-    // device-token: CLI polls the token endpoint ~every 5s (interval); 12/min covers
-    // honest polling with headroom and still caps scripted abuse.
-    options.AddPolicy("device-token", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 12,
-            QueueLimit = 0,
-        });
-    });
-    // device-approve: authed consent action; one submit per CLI session. 10/min per IP.
-    options.AddPolicy("device-approve", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 10,
-            QueueLimit = 0,
-        });
-    });
-    // Per-IP partition — bot with one IP can't exhaust the limit for everyone.
-    // 3 guest-creates per 5min per IP: covers legit shared-WiFi cases, blocks scripted abuse.
-    // ForwardedHeaders runs before RateLimiter in the pipeline, so RemoteIpAddress is the real client.
-    options.AddPolicy("guest-session", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(5),
-            PermitLimit = 3,
-            QueueLimit = 0,
-        });
-    });
-    // "Send to TextStack" web clip receiver — per-IP cap. Each clip queues an
-    // ingestion job + stores HTML, so this blocks scripted bulk-clipping while
-    // staying generous for a human saving a handful of articles in a session.
-    options.AddPolicy("clip", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 20,
-            QueueLimit = 0,
-        });
-    });
-    // User book upload — per-IP cap, mirrors the clip zone. Uploads are heavier
-    // (file ingestion) so the same conservative bucket applies.
-    options.AddPolicy("user-upload", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 20,
-            QueueLimit = 0,
-        });
-    });
-    // TTS synthesis — per-IP cap. Generous enough for bursty vocab-review /
-    // reader-tap usage (~2/s average, tolerates ~20-req bursts via window
-    // timing), but blocks scripted abuse hammering the upstream Bing WS.
-    // ETag + server+client cache mean most requests are cheap; this limit
-    // primarily protects synthesis (uncached) throughput.
-    options.AddPolicy("tts", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 120,
-            QueueLimit = 0,
-        });
-    });
-    // TTS voices list — cheap (served from 24h server cache), typically
-    // called once per session at voice-picker mount. Separate bucket with
-    // a much higher cap so a user burning through synthesis budget can
-    // still open the voice picker.
-    options.AddPolicy("tts-voices", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 600,
-            QueueLimit = 0,
-        });
-    });
-    // Translation — per-IP. OpenAI is the upstream cost; users
-    // typically call 1-3 times per reading session via SelectionToolbar.
-    // 30/min is generous for normal UX, blocks scripted scraping of the
-    // translation service.
-    options.AddPolicy("translate", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 30,
-            QueueLimit = 0,
-        });
-    });
-    // Dictionary lookup — per-IP. Proxies Free Dictionary API (cheap,
-    // public). Users tap words rapidly while reading; 60/min fits the
-    // natural pace and still caps scripted abuse.
-    options.AddPolicy("dictionary", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 60,
-            QueueLimit = 0,
-        });
-    });
-    // /explain — Claude/OpenAI-backed contextual explanation. Paid API
-    // call per miss (cache fronts it). 20/min per IP is plenty for
-    // active reading; anything higher smells like scripting.
-    options.AddPolicy("explain", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 20,
-            QueueLimit = 0,
-        });
-    });
-    // Hybrid catalog search (AI-057): semantic=true embeds the query (one paid OpenAI embedding
-    // call per request) before the $0 pgvector scan, so it gets its own per-IP throttle. CRITICAL:
-    // this policy is a NO-OP unless `semantic` is truthy — the pure-FTS path (semantic absent/false)
-    // consumes no partition and stays completely unthrottled (zero new cost/latency).
-    options.AddPolicy("search-semantic", httpContext =>
-    {
-        var semantic = httpContext.Request.Query["semantic"].ToString();
-        var isSemantic = semantic.Equals("true", StringComparison.OrdinalIgnoreCase)
-                         || semantic == "1";
-        if (!isSemantic)
-            return RateLimitPartition.GetNoLimiter("search-fts");
-
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter("semantic:" + ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 20,
-            QueueLimit = 0,
-        });
-    });
-    // "Ask this book" (RAG) — one LLM call per request, per-user reading. 30/min per IP is
-    // generous for genuine use and caps scripted abuse.
-    options.AddPolicy("rag.ask", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 30,
-            QueueLimit = 0,
-        });
-    });
-    // On-demand "Ask this book" index trigger (Phase 1): per-IP cap so a user can't mass-index
-    // the whole catalog. ~20/hour — generous for legit "index this book then poll" flows.
-    options.AddPolicy("rag.index", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromHours(1),
-            PermitLimit = 20,
-            QueueLimit = 0,
-        });
-    });
-    // Study Buddy agent (AI-037): each run is several LLM calls, so a tighter per-IP limit.
-    options.AddPolicy("studybuddy", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 8,
-            QueueLimit = 0,
-        });
-    });
-    // Librarian agent (AI-Agent-3): each run is several LLM calls + maybe external HTTP, so a tight per-IP cap.
-    options.AddPolicy("librarian", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 8,
-            QueueLimit = 0,
-        });
-    });
-    // Learning Tutor agent (AI-Agent-2): each planning turn is several LLM calls + DB reads, so a tight per-IP
-    // cap. Mirrors the librarian policy shape.
-    options.AddPolicy("tutor", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 8,
-            QueueLimit = 0,
-        });
-    });
-    // AutoPublish crew (AI-042): an admin generate is TWO 4-stage crews = 8 LLM calls, so a tight per-IP cap.
-    // Mirrors the studybuddy policy shape; it sits behind admin auth too, this is just runaway protection.
-    options.AddPolicy("autopublish.crew", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 4,
-            QueueLimit = 0,
-        });
-    });
-    // SEO crew (AI-043): one 4-stage crew = 4 LLM calls per admin generate. Same tight per-IP cap as
-    // autopublish.crew; sits behind admin auth too, this is just runaway protection.
-    options.AddPolicy("seo.crew", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(1),
-            PermitLimit = 4,
-            QueueLimit = 0,
-        });
-    });
-    // Account deletion — destructive and irreversible (GDPR hard delete). A user
-    // never needs to call this more than once, so a tight per-IP cap blocks abuse
-    // (e.g. scripted churn against the cascade delete) while staying out of the way
-    // of a legitimate retry after a transient failure.
-    options.AddPolicy("account-delete", httpContext =>
-    {
-        var ip = httpContext.Connection.RemoteIpAddress?.ToString() ?? "unknown";
-        return RateLimitPartition.GetFixedWindowLimiter(ip, _ => new FixedWindowRateLimiterOptions
-        {
-            Window = TimeSpan.FromMinutes(5),
-            PermitLimit = 3,
-            QueueLimit = 0,
-        });
-    });
-    options.RejectionStatusCode = StatusCodes.Status429TooManyRequests;
-    // Emit Retry-After so clients can back off intelligently instead of
-    // hammering in a tight retry loop. RateLimiter exposes the metadata
-    // via lease.TryGetMetadata when available.
-    options.OnRejected = (context, ct) =>
-    {
-        if (context.Lease.TryGetMetadata(MetadataName.RetryAfter, out var retryAfter))
-        {
-            var seconds = (int)Math.Ceiling(retryAfter.TotalSeconds);
-            context.HttpContext.Response.Headers.RetryAfter = seconds.ToString();
-        }
-        return ValueTask.CompletedTask;
-    };
-});
+// Persistence, search, RAG, product services, workers, and rate-limiting.
+builder.Services
+    .AddTextStackPersistence(connectionString, builder.Configuration)
+    .AddTextStackSearchStack(connectionString, builder.Configuration)
+    .AddTextStackRag(connectionString)
+    .AddTextStackContentServices(builder.Configuration)
+    .AddTextStackHostedServices()
+    .AddTextStackRateLimiting();
 
 // Validate required config at startup
 if (!builder.Environment.IsEnvironment("Test"))
